Move SearchableItem to directory
