+++ src/schema/SearchableItem/index.ts
