+++ src/schema/search/index.ts
@@ -14,7 +14,7 @@ import { pageable } from "relay-cursor-paging"
 import { connectionWithCursorInfo } from "schema/fields/pagination"
 import { createPageCursors } from "schema/fields/pagination"
 import { convertConnectionArgsToGravityArgs } from "lib/helpers"
-import { SearchableItem } from "schema/searchableItem"
+import { SearchableItem } from "schema/SearchableItem"
 import { Searchable } from "schema/searchable"
 import { SearchEntity } from "./SearchEntity"
 import { ResolverContext } from "types/graphql"
